requirements.txt
asgiref==3.4.1
Django==3.2.7
django-debug-toolbar==3.2.2
django-environ==0.7.0
django-filter==21.1
django-social-share==2.2.1
lazy-object-proxy==1.6.0
Pillow==8.3.2
psycopg2-binary==2.9.1
python-dotenv==0.19.0
pytils==0.3
pytz==2021.1
sqlparse==0.4.2
toml==0.10.2
typing-extensions==3.10.0.2
wrapt==1.12.1

aiogram~=2.15